Prob stat and solution code - no solution pdf
The research sections in the problem statement of this session are left as an exercise for the learner. We only focus on the code to get some values for specs and dimensions to use in the ADCS. We need the radius of the reaction wheel to build it as the material and other dimensions are fixed. We need the moment of inertia of the reaction wheel to get the factor of safety for storing the angular momentum of the craft in the reaction wheel. Similarly, we need the length of wire, number of coils, and length of core to build the magnetorquer. We need the magnetic moment generated by the magnetorquer to get the factor safety compared to the magnetic moment we originally needed.